transferred classes in Core Package
