--- app/src/main/java/at/aau/risiko/cards/CardService.java
@@ -1,10 +1,0 @@
-package at.aau.risiko.cards;
-
-import java.util.ArrayList;
-import java.util.Random;
-
-public class CardService {
-
-
-
-}
